# publisher/docs/swagger.yaml
basePath: /
definitions:
  controllers.sent:
    properties:
      message:
        type: string
    type: object
  model.Attachment:
    properties:
      confirmedUpload:
        type: boolean
      contentType:
        type: string
      createdAt:
        type: string
      id:
        type: string
      minioName:
        type: string
      name:
        type: string
      size:
        type: integer
      userId:
        type: string
    type: object
  model.AttachmentPartial:
    properties:
      contentType:
        type: string
      name:
        type: string
      size:
        minimum: 1
        type: integer
    required:
    - contentType
    - name
    - size
    type: object
  model.AttachmentURL:
    properties:
      formData:
        additionalProperties:
          type: string
        type: object
      id:
        type: string
      minioName:
        type: string
      url:
        type: string
    type: object
...
  model.EmailList:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      deletedAt:
        type: string
      deletedBy:
        type: string
      description:
        type: string
      emailAlias:
        type: string
      emails:
        additionalProperties:
          type: string
        description: Email uses map[id] to have the possibility to remove by ID without
          revealing which email will be removed
        type: object
      id:
        type: string
      name:
        type: string
    type: object
...
  model.EmailListPartial:
    properties:
      description:
        type: string
      emailAlias:
        type: string
      emails:
        items:
          type: string
        minItems: 1
        type: array
      name:
        type: string
    required:
    - description
    - emailAlias
    - emails
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is an api that publishes emails in RabbitMQ.
  title: Publisher Emails
  version: "1.0"
paths:
  /email/attachment:
    get:
      consumes:
      - application/json
      description: Get all user attachments.
      produces:
      - application/json
      responses:
        "200":
          description: all attachments
          schema:
            items:
              $ref: '#/definitions/model.Attachment'
            type: array
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Get user attachments
      tags:
      - attachment
    post:
      consumes:
      - application/json
      description: Create a upload attachment url.
      parameters:
      - description: attachment params
        in: body
        name: attachment
        required: true
        schema:
          $ref: '#/definitions/model.AttachmentPartial'
      produces:
      - application/json
      responses:
        "200":
          description: create attachment successfully
          schema:
            $ref: '#/definitions/model.AttachmentURL'
        "400":
          description: an invalid attachment param was sent
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Creating attachment
      tags:
      - attachment
  /email/attachment/{id}:
    get:
      consumes:
      - application/json
      description: Get a download attachment url.
      parameters:
      - description: attachment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: attachment url
          schema:
            $ref: '#/definitions/model.AttachmentURL'
        "400":
          description: was sent a invalid attachment ID
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: attachment does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Get attachment url
      tags:
      - attachment
    post:
      consumes:
      - application/json
      description: Refresh a upload attachment url.
      parameters:
      - description: attachment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: attachment url
          schema:
            $ref: '#/definitions/model.AttachmentURL'
        "400":
          description: was sent a invalid attachment ID
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: attachment does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Refresh attachment url
      tags:
      - attachment
  /email/attachment/{id}/confirm:
    post:
      consumes:
      - application/json
      description: Confirm upload.
      parameters:
      - description: attachment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: upload confimed
          schema:
            $ref: '#/definitions/controllers.sent'
        "400":
          description: was sent a invalid attachment ID
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: attachment does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Confirm upload
      tags:
      - attachment
  /email/list:
    get:
      consumes:
      - application/json
      description: Get all user email list.
      produces:
      - application/json
      responses:
        "200":
          description: email list
          schema:
            items:
              $ref: '#/definitions/model.EmailList'
            type: array
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Get all user email list
      tags:
      - emailList
    post:
      consumes:
      - application/json
      description: Create a email list to user.
      parameters:
      - description: email list params
        in: body
        name: emailList
        required: true
        schema:
          $ref: '#/definitions/model.EmailListPartial'
      produces:
      - application/json
      responses:
        "201":
          description: create email list successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "400":
          description: an invalid email list param was sent
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "409":
          description: email list already exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Creating email list
      tags:
      - emailList
  /email/list/{name}:
    delete:
      consumes:
      - application/json
      description: Delete an email list.
      parameters:
      - description: email list name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: delete email list successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: email list does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Delete email list
      tags:
      - emailList
    get:
      consumes:
      - application/json
      description: Get a user email list.
      parameters:
      - description: email list name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: email list
          schema:
            $ref: '#/definitions/model.EmailList'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: email list not found
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Get a user email list
      tags:
      - emailList
    put:
      consumes:
      - application/json
      description: Update information for an email list.
      parameters:
      - description: email list name
        in: path
        name: name
        required: true
        type: string
      - description: email list info
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/model.EmailListInfo'
      produces:
      - application/json
      responses:
        "200":
          description: update email list successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "400":
          description: an invalid email list param was sent
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: email list does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "409":
          description: email list already exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Update info email list
      tags:
      - emailList
  /email/list/{name}/add:
    post:
      consumes:
      - application/json
      description: Add emails to email list.
      parameters:
      - description: email list name
        in: path
        name: name
        required: true
        type: string
      - description: emails
        in: body
        name: emails
        required: true
        schema:
          $ref: '#/definitions/model.EmailListEmails'
      produces:
      - application/json
      responses:
        "200":
          description: update email list successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "400":
          description: an invalid email param was sent
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: email list does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Add emails to email list
      tags:
      - emailList
  /email/list/{name}/remove:
    delete:
      consumes:
      - application/json
      description: Remove emails to email list.
      parameters:
      - description: email list name
        in: path
        name: name
        required: true
        type: string
      - description: emails
        in: body
        name: emails
        required: true
        schema:
          $ref: '#/definitions/model.EmailListEmails'
      produces:
      - application/json
      responses:
        "200":
          description: update email list successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "400":
          description: an invalid email param was sent
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: email list does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Remove emails to email list
      tags:
      - emailList
  /email/list/{user_id}/{name}/{email_id}:
    delete:
      consumes:
      - application/json
      description: Remove emails to email list.
      parameters:
      - description: user id from emailist
        in: path
        name: user_id
        required: true
        type: string
      - description: email list name
        in: path
        name: name
        required: true
        type: string
      - description: email id from emailist
        in: path
        name: email_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: update email list successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: email list does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Remove email to email list
      tags:
      - emailList
  /email/queue:
    get:
      consumes:
      - application/json
      description: Get all RabbitMQ queues.
      produces:
      - application/json
      responses:
        "200":
          description: all queues
          schema:
            items:
              $ref: '#/definitions/model.Queue'
            type: array
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Get queues
      tags:
      - queue
    post:
      consumes:
      - application/json
      description: Create a RabbitMQ queue with DLX.
      parameters:
      - description: queue params
        in: body
        name: queue
        required: true
        schema:
          $ref: '#/definitions/model.QueuePartial'
      produces:
      - application/json
      responses:
        "201":
          description: create queue successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "400":
          description: an invalid queue param was sent
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "403":
          description: current user is not admin
          schema:
            $ref: '#/definitions/controllers.sent'
        "409":
          description: queue already exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Creating queue
      tags:
      - queue
  /email/queue/{name}:
    delete:
      consumes:
      - application/json
      description: Delete a queue with DLX.
      parameters:
      - description: queue name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: queue deleted
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "403":
          description: current user is not admin
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: queue does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Delete queues
      tags:
      - queue
  /email/queue/{name}/send:
    post:
      consumes:
      - application/json
      description: Sends an email to the RabbitMQ queue.
      parameters:
      - description: queue name
        in: path
        name: name
        required: true
        type: string
      - description: email
        in: body
        name: queue
        required: true
        schema:
          $ref: '#/definitions/model.Email'
      produces:
      - application/json
      responses:
        "200":
          description: email sent successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "400":
          description: an invalid email param was sent
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "403":
          description: current user is not admin
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: queue does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Sends email
      tags:
      - queue
  /email/template:
    get:
      consumes:
      - application/json
      description: Get all user templates.
      produces:
      - application/json
      responses:
        "200":
          description: all user templates
          schema:
            items:
              $ref: '#/definitions/model.Template'
            type: array
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Get user template
      tags:
      - template
    post:
      consumes:
      - application/json
      description: Create a email template.
      parameters:
      - description: template params
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/model.TemplatePartial'
      produces:
      - application/json
      responses:
        "201":
          description: create template successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "400":
          description: an invalid template param was sent
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "409":
          description: template name already exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Creating template
      tags:
      - template
  /email/template/{name}:
    delete:
      consumes:
      - application/json
      description: Delete a email template.
      parameters:
      - description: template name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: template deleted
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: template does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Delete template
      tags:
      - template
    get:
      consumes:
      - application/json
      description: Get a email template.
      parameters:
      - description: template name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: all templates
          schema:
            $ref: '#/definitions/model.Template'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: template does not exist
          schema:
            items:
              $ref: '#/definitions/controllers.sent'
            type: array
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Get template
      tags:
      - template
    put:
      consumes:
      - application/json
      description: Update a email template.
      parameters:
      - description: template name
        in: path
        name: name
        required: true
        type: string
      - description: template params
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/model.TemplatePartial'
      produces:
      - application/json
      responses:
        "200":
          description: template updated
          schema:
            $ref: '#/definitions/controllers.sent'
        "400":
          description: an invalid template param was sent
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: template does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Update template
      tags:
      - template
  /email/template/all:
    get:
      consumes:
      - application/json
      description: Get all email templates.
      produces:
      - application/json
      responses:
        "200":
          description: all templates
          schema:
            items:
              $ref: '#/definitions/model.Template'
            type: array
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Get templates
      tags:
      - template
  /user:
    delete:
      consumes:
      - application/json
      description: Delete current user.
      produces:
      - application/json
      responses:
        "200":
          description: user deleted
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "403":
          description: current user is protected
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: user does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Delete user
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Get current user informations.
      produces:
      - application/json
      responses:
        "200":
          description: user informations
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: user does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Get user
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Create a user in application.
      parameters:
      - description: user params
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserPartial'
      produces:
      - application/json
      responses:
        "201":
          description: user created successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "400":
          description: an invalid user param was sent
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "403":
          description: current user is not admin
          schema:
            $ref: '#/definitions/controllers.sent'
        "409":
          description: user already exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Create user
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Update user informatios.
      parameters:
      - description: user params
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserPartial'
      produces:
      - application/json
      responses:
        "200":
          description: user updated
          schema:
            $ref: '#/definitions/controllers.sent'
        "400":
          description: an invalid user param was sent
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: user does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Update user
      tags:
      - user
...
  /user/all:
    get:
      consumes:
      - application/json
      description: Get all users informations.
      produces:
      - application/json
      responses:
        "200":
          description: user informations
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "403":
          description: current user is not admin
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Get all users
      tags:
      - admin
...
swagger: "2.0"